requirements.txt
async-timeout==4.0.3
asyncpg==0.30.0
python-dotenv==1.0.1
loguru==0.7.3